% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_trabalho.R
\name{cria_gitpage}
\alias{cria_gitpage}
\title{Cria um novo projeto de pagina}
\usage{
cria_gitpage(filename = "my-gitpage")
}
\arguments{
\item{filename}{}
}
\description{
Cria um novo projeto de pagina
}
