% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_trabalho.R
\name{cria_trabalho}
\alias{cria_trabalho}
\title{Cria um novo arquivo .rmd de um trabalho}
\usage{
cria_trabalho(filename = "Francisco_Rosa_Dias_de_Miranda_Teste08.rmd")
}
\arguments{
\item{filename}{Nome do arquivo}
}
\description{
Cria um novo arquivo .rmd de um trabalho
}
